+++ packages/node/src/login/oidc/ClientRegistrar.ts
@@ -35,6 +35,7 @@ import {
   ConfigurationError,
   determineSigningAlg,
   PREFERRED_SIGNING_ALG,
+  isKnownClientType,
 } from "@inrupt/solid-client-authn-core";
 import type { Client } from "openid-client";
 import { Issuer } from "openid-client";
@@ -86,6 +87,7 @@ export default class ClientRegistrar implements IClientRegistrar {
       storedClientSecret,
       storedClientName,
       storedIdTokenSignedResponseAlg,
+      storedClientType,
     ] = await Promise.all([
       this.storageUtility.getForUser(options.sessionId, "clientId"),
       this.storageUtility.getForUser(options.sessionId, "clientSecret"),
@@ -94,17 +96,22 @@ export default class ClientRegistrar implements IClientRegistrar {
         options.sessionId,
         "idTokenSignedResponseAlg",
       ),
+      this.storageUtility.getForUser(options.sessionId, "clientType"),
     ]);
-    if (storedClientId) {
+
+    if (storedClientId !== undefined && isKnownClientType(storedClientType)) {
+      if (storedClientType === "static" && storedClientSecret === undefined) {
+        throw new Error("Missing static client secret in storage.");
+      }
       return {
         clientId: storedClientId,
         clientSecret: storedClientSecret,
-        clientName: storedClientName as string | undefined,
+        clientName: storedClientName,
         idTokenSignedResponseAlg:
           storedIdTokenSignedResponseAlg ??
           negotiateClientSigningAlg(issuerConfig, PREFERRED_SIGNING_ALG),
-        clientType: "dynamic",
-      };
+        clientType: storedClientType,
+      } as IClient;
     }
 
     // TODO: It would be more efficient to only issue a single request (see IssuerConfigFetcher)
@@ -135,21 +142,27 @@ export default class ClientRegistrar implements IClientRegistrar {
       grant_types: ["authorization_code", "refresh_token"],
     });
 
-    const infoToSave: Record<string, string> = {
+    const persistedClientMetadata: Record<string, string> = {
       clientId: registeredClient.metadata.client_id,
       idTokenSignedResponseAlg:
         registeredClient.metadata.id_token_signed_response_alg ?? signingAlg,
+      clientType: "dynamic",
     };
-    if (registeredClient.metadata.client_secret) {
-      infoToSave.clientSecret = registeredClient.metadata.client_secret;
+    if (registeredClient.metadata.client_secret !== undefined) {
+      persistedClientMetadata.clientSecret =
+        registeredClient.metadata.client_secret;
     }
-    await this.storageUtility.setForUser(options.sessionId, infoToSave);
+
+    await this.storageUtility.setForUser(
+      options.sessionId,
+      persistedClientMetadata,
+    );
     return {
-      clientId: registeredClient.metadata.client_id,
-      clientSecret: registeredClient.metadata.client_secret,
-      clientName: registeredClient.metadata.client_name as string | undefined,
+      clientId: persistedClientMetadata.clientId,
+      clientSecret: persistedClientMetadata.clientSecret,
       idTokenSignedResponseAlg:
-        registeredClient.metadata.id_token_signed_response_alg ?? signingAlg,
+        persistedClientMetadata.idTokenSignedResponseAlg,
+      clientName: registeredClient.metadata.client_name as string | undefined,
       clientType: "dynamic",
     };
   }