+++ packages/node/src/login/oidc/incomingRedirectHandler/AuthCodeRedirectHandler.ts
@@ -126,9 +126,10 @@ export class AuthCodeRedirectHandler implements IIncomingRedirectHandler {
     const client = new issuer.Client({
       client_id: clientInfo.clientId,
       client_secret: clientInfo.clientSecret,
-      token_endpoint_auth_method: clientInfo.clientSecret
-        ? "client_secret_basic"
-        : "none",
+      token_endpoint_auth_method:
+        typeof clientInfo.clientSecret === "undefined"
+          ? "none"
+          : "client_secret_basic",
       id_token_signed_response_alg: clientInfo.idTokenSignedResponseAlg,
     });